requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cloudinary==1.28.0
Django==4.0
django-filter==21.1
gunicorn
Pillow==9.0.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise
